# packages/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /network/jobs:
    get:
      operationId: listNetworkJobs
      tags: [network]
      summary: List public network jobs
      description: Public technical-beta job records, newest first. Native provider reports are not exposed.
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [completed, rejected]
      responses:
        "200":
          description: Public job records
          content:
            application/json:
              schema:
                type: array
                maxItems: 100
                items:
                  $ref: "#/components/schemas/PublicNetworkJob"
  /network/jobs/{jobId}:
    get:
      operationId: getNetworkJob
      tags: [network]
      summary: Get a public network job
      description: Public technical-beta job detail. A result is returned only when it is already stored on the job.
      parameters:
        - $ref: "#/components/parameters/JobId"
      responses:
        "200":
          description: Public job detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublicNetworkJob"
        "404":
          description: Job not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /receipts/{jobId}:
    get:
      operationId: getJobReceipt
      tags: [receipts]
      summary: Get a server-issued technical-beta receipt
      description: Returns a deterministic receipt for a completed Native Node job.
      parameters:
        - $ref: "#/components/parameters/JobId"
      responses:
        "200":
          description: Server-issued receipt
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobReceipt"
        "404":
          description: Completed job receipt not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /verifications:
    post:
      operationId: createVerification
      tags: [verifications]
      summary: Canonically verify a completed job
      description: Recomputes a completed Native Node job with the server canonical engine and records the comparison.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VerificationInput"
      responses:
        "201":
          description: Verification event recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VerificationEvent"
        "400":
          description: Invalid verification request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Completed job not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    get:
      operationId: listVerifications
      tags: [verifications]
      summary: List verification events
      description: Returns the newest 100 server canonical verification events.
      responses:
        "200":
          description: Verification events
          content:
            application/json:
              schema:
                type: array
                maxItems: 100
                items:
                  $ref: "#/components/schemas/VerificationEvent"
components:
  parameters:
...
    JobId:
      name: jobId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  schemas:
...
    PublicNetworkJob:
      type: object
      description: Publicly visible redacted job status. Inputs, results, leases, and agent attribution are private.
      additionalProperties: false
      required: [id, providerId, status, cycles, createdAt, completedAt]
      properties:
        id: { type: string, format: uuid }
        providerId: { type: string, format: uuid }
        status: { type: string, enum: [completed, rejected] }
        cycles: { type: integer, minimum: 1, maximum: 32 }
        createdAt: { type: string, format: date-time }
        completedAt: { type: ["string", "null"], format: date-time }
...
    JobReceipt:
      type: object
      description: A server-issued technical-beta receipt for Native Node execution.
      additionalProperties: false
      required: [receiptId, jobId, providerId, agentId, agentPolicyHash, agentPolicyVersion, agentOwnerWalletAddress, engine, digestAlgorithm, resultDigest, inputs, cycles, createdAt, assignedAt, completedAt, verificationMethod, verificationStatus]
      properties:
        receiptId: { type: string, minLength: 1, maxLength: 128 }
        jobId: { type: string, format: uuid }
        providerId: { type: string, format: uuid }
        agentId: { type: ["string", "null"], format: uuid }
        agentPolicyHash: { type: ["string", "null"], pattern: "^[a-f0-9]{64}$" }
        agentPolicyVersion: { type: ["integer", "null"], minimum: 1 }
        agentOwnerWalletAddress: { type: ["string", "null"], pattern: "^0x[a-fA-F0-9]{40}$" }
        engine: { type: string }
        digestAlgorithm: { type: string, enum: [SHA-256] }
        resultDigest: { type: string, pattern: "^[a-f0-9]{64}$" }
        inputs:
          type: array
          minItems: 1
          maxItems: 8
          items: { type: integer, minimum: 0, maximum: 255 }
        cycles: { type: integer, minimum: 1, maximum: 32 }
        createdAt: { type: string, format: date-time }
        assignedAt: { type: ["string", "null"], format: date-time }
        completedAt: { type: string, format: date-time }
        verificationMethod: { type: string, enum: [server-canonical-recomputation] }
        verificationStatus: { type: string, enum: [accepted] }
    VerificationInput:
      type: object
      additionalProperties: false
      required: [jobId]
      properties:
        jobId: { type: string, format: uuid }
    VerificationEvent:
      type: object
      description: A persisted server canonical recomputation event for Native Node execution.
      additionalProperties: false
      required: [id, jobId, status, digest, checkedAt]
      properties:
        id: { type: string, format: uuid }
        jobId: { type: string, format: uuid }
        status: { type: string, enum: [match, mismatch] }
        digest: { type: string, pattern: "^[a-f0-9]{64}$" }
        checkedAt: { type: string, format: date-time }
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error